Fix warnings associated with 64-bit/32-bit compiler.

+++ Classes/BDDynamicGridViewController.m
@@ -186,7 +186,7 @@
             _rowInfos = [_rowInfos arrayByAddingObject:ri];
         }
         ri.isLastCell = YES;
-        NSAssert(accumNumOfViews == self.delegate.numberOfViews, @"wrong accum %u ", ri.accumulatedViews);
+        NSAssert(accumNumOfViews == self.delegate.numberOfViews, @"wrong accum %lu ", (unsigned long)ri.accumulatedViews);
     }else{
         _rowInfos = [self.delegate customLayout];
     }
@@ -218,7 +218,7 @@
             _rowInfos = [_rowInfos arrayByAddingObject:ri];
         }
         ri.isLastCell = YES;
-        NSAssert(accumNumOfViews == self.delegate.numberOfViews, @"wrong accum %u ", ri.accumulatedViews);
+        NSAssert(accumNumOfViews == self.delegate.numberOfViews, @"wrong accum %lu ", (unsigned long)ri.accumulatedViews);
         [_tableView reloadData];
     }
 }
@@ -280,7 +280,7 @@
 {
     BDRowInfo *ri = [_rowInfos objectAtIndex:indexPath.row];
     static NSString *CellIdentifier = @"Cell";
-    BDDynamicGridCell *cell = [tableView dequeueReusableCellWithIdentifier:[CellIdentifier stringByAppendingFormat:@"_viewCount%d", ri.viewsPerCell]];
+    BDDynamicGridCell *cell = [tableView dequeueReusableCellWithIdentifier:[CellIdentifier stringByAppendingFormat:@"_viewCount%lu", (unsigned long)ri.viewsPerCell]];
     
     if (!cell) {
         cell = [[BDDynamicGridCell alloc] initWithLayoutStyle:BDDynamicGridCellLayoutStyleFill
